+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 1.6 - Unreleased
 * Block: Added a method to encode contents in Base64.
 * TlsRequest: Avoid possible crash when reusing TLS sessions.
+* TlsRequest: Added a method to query the SSL library name (OpenSSL/LibreSSL).
 
 ## 1.5 - 2022-11-25
 * RegExp: PCRE2 can be used as an alternative to the older PCRE.